Renamed TagAttribute to GameobjectTagAttribute
