+++ Runtime/Mapbox/BaseModule/Utilities/Attributes/GameobjectTagAttribute.cs
@@ -5,5 +5,5 @@ namespace Mapbox.BaseModule.Utilities.Attributes
 {
 
     [AttributeUsage(AttributeTargets.Field, AllowMultiple = false, Inherited = true)]
-    public class TagAttribute : PropertyAttribute {}
+    public class GameobjectTagAttribute : PropertyAttribute {}
 }
